GreenBite-backend/requirements.txt
asgiref==3.11.0
Django>=5.0,<6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
idna==3.11
Markdown==3.10
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
sqlparse==0.5.4
urllib3==2.6.0
psycopg2-binary==2.9.6
openai==2.14.0
httpx==0.27.2
drf-yasg
numpy==1.26.4
celery==5.3.6
redis==5.0.1
django-celery-results==2.5.1
django-redis==5.2.0
pgvector==0.4.2